[#23] refactor: divide domain
