+++ data/lib/local_source/local_source.dart
@@ -1,4 +1,5 @@
 import 'package:data/model/schema/schema.dart';
 import 'package:isar/isar.dart';
+import 'package:domain/domain.dart' as domain;
 
 part 'recipe_isar.dart';
