+++ data/lib/local_source/recipe_isar.dart
@@ -1,36 +1,44 @@
 part of 'local_source.dart';
 
-class RecipeIsar implements RecipeSource {
+class RecipeIsar implements domain.RecipeSource {
   final _openIsar = Isar.open([
     IngredientSchema,
     RecipeSchema,
   ]);
 
-  Future<int> addRecipe(Recipe recipe) => _upsertRecipe(recipe);
+  @override
+  Future<int> addRecipe(domain.Recipe recipe) => _upsertRecipe(recipe);
 
-  Future<Recipe?> getRecipe(Id id) async {
+  @override
+  Future<domain.Recipe?> getRecipe(Id id) async {
     final isar = await _openIsar;
     return isar.recipes.getSync(id);
   }
 
-  Future<int> updateRecipe(Recipe recipe) => _upsertRecipe(recipe);
+  @override
+  Future<int> updateRecipe(domain.Recipe recipe) => _upsertRecipe(recipe);
 
+  @override
   Future<bool> deleteRecipe(Id id) async {
     final isar = await _openIsar;
     return isar.recipes.deleteSync(id);
   }
 
-  Future<int> addIngredient(Ingredient ingredient) =>
+  @override
+  Future<int> addIngredient(domain.Ingredient ingredient) =>
       _upsertIngredient(ingredient);
 
-  Future<Ingredient?> getIngredient(Id id) async {
+  @override
+  Future<domain.Ingredient?> getIngredient(Id id) async {
     final isar = await _openIsar;
     return isar.ingredients.getSync(id);
   }
 
-  Future<int> updateIngredient(Ingredient ingredient) =>
+  @override
+  Future<int> updateIngredient(domain.Ingredient ingredient) =>
       _upsertIngredient(ingredient);
 
+  @override
   Future<bool> deleteIngredient(Id id) async {
     final isar = await _openIsar;
     return isar.ingredients.deleteSync(id);
