+++ domain/lib/data_source/data_source.dart
@@ -1,0 +1,1 @@
+export 'recipe_source.dart';
