+++ domain/lib/domain.dart
@@ -1,0 +1,2 @@
+export 'data_source/data_source.dart';
+export 'model/model.dart';
