+++ domain/lib/model/amount/amount.dart
