+++ domain/lib/model/amount/mass/gram.dart
