+++ domain/lib/model/amount/mass/kilogram.dart
