+++ domain/lib/model/amount/mass/mass.dart
@@ -1,4 +1,4 @@
-import 'package:cooking_calulator/model/util/util.dart';
+import 'package:domain/model/util/util.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 import '../amount.dart';
