+++ domain/lib/model/amount/volume/cubic_centimeter.dart
