+++ domain/lib/model/ingredient.dart
