+++ domain/lib/model/model.dart
