+++ domain/lib/model/temperature/celsius.dart
