+++ domain/lib/model/temperature/fahrenheit.dart
