+++ domain/lib/model/temperature/temperature.dart
