+++ domain/pubspec.yaml
@@ -1,0 +1,11 @@
+name: domain
+description: A sample command-line application.
+version: 0.0.0
+
+environment:
+  sdk: '>=2.17.1 <3.0.0'
+
+dev_dependencies:
+  lints: ^2.0.0
+  test: ^1.16.0
+dependencies: {freezed: ^2.2.0, freezed_annotation: ^2.2.0}
