+++ lib/page/recipe/recipe_edit_page.dart
@@ -1,5 +1,5 @@
-import 'package:cooking_calulator/model/model.dart';
 import 'package:cooking_calulator/widget/widget.dart';
+import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
